move fixes for worldCopyJump to separate module
